requirements.txt
docx==0.2.4
fitz==0.0.1.dev2
numpy==1.26.4 
protobuf==6.31.1
python_docx==1.2.0
scikit_learn==1.7.1
streamlit==1.37.1
google-generativeai==0.8.5 
